20240503 수업

+++ src/main/java/com/prj1/controller/MemberController.java
@@ -4,6 +4,7 @@ package com.prj1.controller;
 import com.prj1.domain.Member;
 import com.prj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,7 +29,7 @@ public class MemberController {
     @PostMapping("sign-up")
     public String signUpPost(Member member) {
         service.signUp(member);
-        return "redirect:/";
+        return "redirect:/member/login";
     }
 
     @GetMapping("list")
@@ -44,9 +45,11 @@ public class MemberController {
     }
 
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
-        return "redirect:/member/sign-up";
+    public String remove(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
+        return "redirect:/logout";
     }
 
     @GetMapping("modify")
@@ -55,8 +58,10 @@ public class MemberController {
     }
 
     @PostMapping("modify")
-    public String modifyPost(Member member, RedirectAttributes rttr) {
-        service.modify(member);
+    public String modifyPost(Member member, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(member.getId(), authentication)) {
+            service.modify(member);
+        }
 
         rttr.addAttribute("id", member.getId());
         return "redirect:/member";
@@ -67,5 +72,10 @@ public class MemberController {
     public String emailCheck(String email) {
         String message = service.emailCheck(email);
         return message;
+    }
+
+    @GetMapping("login")
+    public String loginForm() {
+        return "member/login";
     }
 }
